Lesson 6: created a Profile class and moved this class and the TodoItem and TodoList classes to separate files
