+++ Programs/TodoList/Program.cs
@@ -2,176 +2,37 @@
 
 namespace Todolist
 {
-	public class TodoItem
-	{
-		public string Text { get; private set; }
-		public bool IsDone { get; private set; }
-		public DateTime LastUpdate { get; private set; }
-
-		public TodoItem(string text)
-		{
-			Text = text;
-			IsDone = false;
-			LastUpdate = DateTime.Now;
-		}
-		public void MarkDone()
-		{
-			IsDone = true;
-			LastUpdate = DateTime.Now;
-		}
-		public void UpdateText(string newText)
-		{
-			Text = newText;
-			LastUpdate = DateTime.Now;
-		}
-		public string GetShortInfo()
-		{
-			string shortText = Text.Length > 30 ? Text.Substring(0, 27) + "..." : Text;
-			string status = IsDone ? "Выполнена" : "Не выполнена";
-			string date = LastUpdate.ToString("dd.MM.yyyy HH:mm");
-
-			return $"{shortText,-30} | {status,-12} | {date}";
-		}
-
-		public string GetFullInfo()
-		{
-			string status = IsDone ? "Выполнена" : "Не выполнена";
-			string date = LastUpdate.ToString("dd.MM.yyyy HH:mm");
-
-			return $"Задача: {Text}\nСтатус: {status}\nПоследнее изменение: {date}";
-		}
-	}
-	public class TodoList
-	{
-		private TodoItem[] items;
-		private int count;
-
-		public TodoList(int initialCapacity = 2)
-		{
-			items = new TodoItem[initialCapacity];
-			count = 0;
-		}
-		public void Add(TodoItem item)
-		{
-			if (count >= items.Length)
-			{
-				IncreaseArray(item);
-			}
-			items[count] = item;
-			count++;
-		}
-
-		public void Delete(int index)
-		{
-			if (index < 0 || index >= count)
-				throw new ArgumentOutOfRangeException(nameof(index), "Неверный индекс");
-
-			for (int i = index; i < count - 1; i++)
-			{
-				items[i] = items[i + 1];
-			}
-			items[count - 1] = null;
-			count--;
-		}
-
-		public void View(bool showIndex, bool showStatus, bool showDate)
-		{
-			if (count == 0)
-			{
-				Console.WriteLine("Список пуст");
-				return;
-			}
-			string header = "";
-			if (showIndex) header += "Индекс".PadRight(8);
-			if (showStatus) header += "Статус".PadRight(12);
-			if (showDate) header += "Дата изменения".PadRight(20);
-			header += "Задача";
-
-			Console.WriteLine(header);
-			Console.WriteLine(new string('-', header.Length));
-
-			for (int i = 0; i < count; i++)
-			{
-				string line = "";
-
-				if (showIndex) line += $"{i + 1}".PadRight(8);
-
-				if (showStatus)
-				{
-					string status = items[i].IsDone ? "Сделано" : "Не сделано";
-					line += status.PadRight(12);
-				}
-
-				if (showDate)
-				{
-					string date = items[i].LastUpdate.ToString("dd.MM.yyyy HH:mm");
-					line += date.PadRight(20);
-				}
-
-				string taskText = items[i].Text?.Replace("\n", " ") ?? "";
-				if (taskText.Length > 30)
-					taskText = taskText.Substring(0, 27) + "...";
-				line += taskText;
-
-				Console.WriteLine(line);
-			}
-		}
-
-		public TodoItem GetItem(int index)
-		{
-			if (index < 0 || index >= count)
-				throw new ArgumentOutOfRangeException(nameof(index), "Неверный индекс");
-
-			return items[index];
-		}
-		public int Count => count;
-
-		private void IncreaseArray(TodoItem item)
-		{
-			int newSize = items.Length * 2;
-			TodoItem[] newArray = new TodoItem[newSize];
-
-			Array.Copy(items, newArray, items.Length);
-			items = newArray;
-
-			Console.WriteLine($"Массив расширен до {newSize} элементов");
-		}
-	}
-
 	class Program
     {
 		static TodoList todoList = new TodoList();
-        static string name = "";
-        static string secondName = "";
-        static int birthYear = 0;
+		static Profile userProfile = new Profile();
 
-        static void Main()
+		static void Main()
         {
             bool isValid = true;
             int currentYear = DateTime.Now.Year;
 
             Console.Write("Работу сделали Приходько и Бочкарёв\n");
             Console.Write("Введите свое имя: ");
-            string? name = Console.ReadLine();
-            Console.Write("Введите свою фамилию: ");
-            string? secondName = Console.ReadLine();
-            Console.Write("Введите свой год рождения: ");
-            int birthYear = 0;
+			userProfile.FirstName = Console.ReadLine();
+			Console.Write("Введите свою фамилию: ");
+			userProfile.LastName = Console.ReadLine();
+			Console.Write("Введите свой год рождения: ");
+			int birthYear = 0;
 
             try
             {
-                birthYear = int.Parse(Console.ReadLine());
+                userProfile.BirthYear = int.Parse(Console.ReadLine());
             }
             catch (Exception)
             {
                 isValid = false;
             }
 
-            if ((isValid == true) && (birthYear <= currentYear))
+            if ((isValid == true) && (userProfile.BirthYear <= currentYear))
             {
-                int age = currentYear - birthYear;
-                Console.WriteLine($"Добавлен пользователь:{name},{secondName},возраст - {age}");
-            }
+                Console.WriteLine($"Добавлен пользователь:{userProfile.GetInfo()}");
+			}
 
             else Console.WriteLine("Неверно введен год рождения");
 
@@ -194,7 +55,7 @@ namespace Todolist
                         ShowHelp();
                         break;
                     case "profile":
-                        ShowProfile(name, secondName, birthYear);
+						ShowProfile();
                         break;
                     case "add":
                         ProcessAddCommand(parts);
@@ -251,17 +112,14 @@ namespace Todolist
 				""";
 			Console.WriteLine(t);
         }
-        static void ShowProfile(string? name, string? secondName, int birthYear)
+        static void ShowProfile()
         {
-            if (string.IsNullOrEmpty(name) || string.IsNullOrEmpty(secondName))
+            if (string.IsNullOrEmpty(userProfile.FirstName) || string.IsNullOrEmpty(userProfile.LastName))
             {
                 Console.WriteLine("Данные пользователя не заполнены");
                 return;
             }
-
-            Console.WriteLine($"Имя: {name}");
-            Console.WriteLine($"Фамилия: {secondName}");
-            Console.WriteLine($"Год рождения: {birthYear}");
+			Console.WriteLine(userProfile.GetInfo());
         }
         static void ProcessAddCommand(string[] parts)
         {
